+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-
-0.0.1
+0.2.0
 --------
-- Initial release including the core feature set  
+- Encrypted body request
+- Encrypted body response
 
 0.1.0
 --------
 - Fix server signature creation
 - Fix config parameters
 - Redesign API
+
+0.0.1
+--------
+- Initial release including the core feature set